[feat] : 데이터호출 함수 분리
