+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import "./App.css";
 import { IOrder } from "./Models";
-import assetOrder from "./assets/order.json";
 import Order from "./components/Order/Order";
+import { fetchMyOrder } from "./api/orderApi";
 
 function App() {
   const [isLoading, setLoading] = useState(false);
@@ -10,34 +10,15 @@ function App() {
 
   useEffect(() => {
     showProgress();
-    fetchMyOrder((json) => {
-      parseOrder(json, (order) => {
-        setOrder(order);
-        hideProgress();
-      });
-    });
+    const getOrder = async () => {
+      await fetchMyOrder((order) => setOrder(order));
+      hideProgress();
+    };
+    getOrder();
   }, []);
 
   const showProgress = () => setLoading(true);
   const hideProgress = () => setLoading(false);
-
-  const fetchMyOrder = (onCompleted: (json: object) => void) => {
-    setTimeout(() => {
-      onCompleted(assetOrder);
-    }, 1000);
-  };
-
-  const parseOrder = (json: any, onCompleted: (order: IOrder) => void) => {
-    setTimeout(() => {
-      const order: IOrder = {
-        id: json.id,
-        orderAt: new Date(json.orderAt),
-        amount: json.amount,
-        shipping: json.shipping,
-      };
-      onCompleted(order);
-    }, 500);
-  };
 
   return (
     <div>
